trimmed jsdom test
